// server/src/utils/enum.ts
export enum HttpMessage {
  SUCCESS = "Thành công",
  FAILED = "Thất bại",
  CREATED = "Thêm dữ liệu thành công",
  UPDATED = "Cập nhật dữ liệu thành công",
  DELETED = "Xóa dữ liệu thành công",
}

export enum HttpStatus {
  SUCCESS = 200,
  CREATED = 201,
  NOT_FOUND = 404,
  BAD_REQUEST = 400,
  ERROR = 500,
}